Extract RevertableTxState to separate file
- Created scratchpad/revertable_tx_state.rs with RevertableTxState and all its impls
- Updated scratchpad/mod.rs to re-export RevertableTxState
- Removed duplicate code from mod.rs
- Code compiles successfully
